Fix for usage with importmaps
The chart.js/auto package does not work with importmaps due to it being in a subfolder of the main package.

Since the `auto` package is really simple and stimulus-components are mainly used in rails apps that now default to importmaps for javascript management, it seems to be sensible to integrate the `auto` package code in this stimulus controller to solve the issue.

An other option would have been to fix the issue in the chart.js repository but it seems more complicated because it's not very clear how they build the `auto` package.